+++ src/telnetd.cpp
@@ -132,6 +132,14 @@ void TelnetService::initial_idle_timeout(unsigned long timeout) {
 	initial_idle_timeout_ = timeout;
 }
 
+unsigned long TelnetService::default_write_timeout() const {
+	return write_timeout_;
+}
+
+void TelnetService::default_write_timeout(unsigned long timeout) {
+	write_timeout_ = timeout;
+}
+
 void TelnetService::loop() {
 	for (auto it = connections_.begin(); it != connections_.end(); ) {
 		if (!it->loop()) {
@@ -155,7 +163,7 @@ void TelnetService::loop() {
 #if UUID_TELNETD_HAVE_WIFICLIENT_REMOTE
 			logger_.info(F("New connection from [%s]:%u accepted"), uuid::printable_to_string(client.remoteIP()).c_str(), client.remotePort());
 #endif
-			connections_.emplace_back(shell_factory_, std::move(client), initial_idle_timeout_);
+			connections_.emplace_back(shell_factory_, std::move(client), initial_idle_timeout_, write_timeout_);
 #if !(UUID_TELNETD_HAVE_WIFICLIENT_REMOTE)
 			logger_.info(F("New connection %p accepted"), &connections_.back());
 #endif
@@ -163,7 +171,7 @@ void TelnetService::loop() {
 	}
 }
 
-TelnetService::Connection::Connection(shell_factory_function &shell_factory, WiFiClient &&client, unsigned long idle_timeout)
+TelnetService::Connection::Connection(shell_factory_function &shell_factory, WiFiClient &&client, unsigned long idle_timeout, unsigned long write_timeout)
 		: client_(std::move(client)), stream_(client_) {
 #if UUID_TELNETD_HAVE_WIFICLIENT_REMOTE
 	// These have to be copied because they're not accessible on closed connections
@@ -181,6 +189,10 @@ TelnetService::Connection::Connection(shell_factory_function &shell_factory, WiF
 	// Disconnect after 30 seconds without a response
 	client_.keepAlive(5, 5, 5);
 #endif
+
+	if (write_timeout > 0) {
+		client_.setTimeout(write_timeout);
+	}
 
 	stream_.start();