+++ docs/changelog.rst
@@ -10,5 +10,7 @@ Added
 * Telnet listening service.
 * Telnet stream handler with option negotiation.
 * Use of TCP keepalives to timeout open connections.
+* Configurable idle timeout.
+* Configurable write timeout.
 
 .. |--| unicode:: U+2013 .. EN DASH